refactor: provider imports and update Supabase client usage
Replaces context and service imports from 'lib/context' and 'lib/auth' with unified imports from the 'providers' directory across multiple files. Updates Supabase client usage to use getSupabaseClient instead of direct imports. Also updates product and category service imports for consistency.
